+++ exporters/metric/prometheus/prometheus_test.go
@@ -34,6 +34,10 @@ import (
 )
 
 func TestPrometheusExporter(t *testing.T) {
+	// #TODO: This test does not adequately verify the type of
+	// prometheus metric exported for all types - for example,
+	// it does not verify that an UpDown- counter is exported
+	// as a gauge. To be improved.
 	exporter, err := prometheus.NewExportPipeline(
 		prometheus.Config{
 			DefaultHistogramBoundaries: []float64{-0.5, 1},
@@ -44,7 +48,7 @@ func TestPrometheusExporter(t *testing.T) {
 	require.NoError(t, err)
 
 	meter := exporter.MeterProvider().Meter("test")
-
+	upDownCounter := otel.Must(meter).NewFloat64UpDownCounter("updowncounter")
 	counter := otel.Must(meter).NewFloat64Counter("counter")
 	valuerecorder := otel.Must(meter).NewFloat64ValueRecorder("valuerecorder")
 
@@ -61,6 +65,12 @@ func TestPrometheusExporter(t *testing.T) {
 
 	expected = append(expected, `counter{A="B",C="D",R="V"} 15.3`)
 
+	_ = otel.Must(meter).NewInt64ValueObserver("intobserver", func(_ context.Context, result otel.Int64ObserverResult) {
+		result.Observe(1, labels...)
+	})
+
+	expected = append(expected, `intobserver{A="B",C="D",R="V"} 1`)
+
 	valuerecorder.Record(ctx, -0.6, labels...)
 	valuerecorder.Record(ctx, -0.4, labels...)
 	valuerecorder.Record(ctx, 0.6, labels...)
@@ -71,6 +81,11 @@ func TestPrometheusExporter(t *testing.T) {
 	expected = append(expected, `valuerecorder_bucket{A="B",C="D",R="V",le="1"} 3`)
 	expected = append(expected, `valuerecorder_count{A="B",C="D",R="V"} 4`)
 	expected = append(expected, `valuerecorder_sum{A="B",C="D",R="V"} 19.6`)
+
+	upDownCounter.Add(ctx, 10, labels...)
+	upDownCounter.Add(ctx, -3.2, labels...)
+
+	expected = append(expected, `updowncounter{A="B",C="D",R="V"} 6.8`)
 
 	compareExport(t, exporter, expected)
 	compareExport(t, exporter, expected)